+++ src/main/kotlin/mashup/spring/elegant/search/config/EnumConverter.kt
@@ -1,6 +1,6 @@
 package mashup.spring.elegant.search.config
 
-import mashup.spring.elegant.search.domain.search.enums.Category
+import mashup.spring.elegant.search.domain.model.Category
 import mashup.spring.elegant.search.domain.search.enums.Feature
 import org.springframework.core.convert.converter.Converter
