+++ src/main/kotlin/mashup/spring/elegant/search/domain/model/Shop.kt
@@ -1,6 +1,5 @@
-package mashup.spring.elegant.search.domain.search
+package mashup.spring.elegant.search.domain.model
 
-import mashup.spring.elegant.search.domain.search.enums.Category
 import org.springframework.data.elasticsearch.annotations.*
 import org.springframework.data.elasticsearch.core.geo.GeoPoint
