+++ src/main/kotlin/mashup/spring/elegant/search/domain/model/ShopField.kt
@@ -1,4 +1,4 @@
-package mashup.spring.elegant.search.domain.search.enums
+package mashup.spring.elegant.search.domain.model
 
 /**
  * Elasticsearch 내 필드 이름 Enum 클래스
