+++ src/main/kotlin/mashup/spring/elegant/search/domain/search/enums/Feature.kt
@@ -1,10 +1,10 @@
 package mashup.spring.elegant.search.domain.search.enums
 
-import mashup.spring.elegant.search.domain.search.enums.ShopField.*
+import mashup.spring.elegant.search.domain.model.ShopField.*
 
 /**
  * Feature 를 다루기 위한 Enum
- * ElasticSearch Query 를 숨긴다.
+ * API 에 대해 ElasticSearch Query 를 숨긴다.
  */
 enum class Feature(
     val field : String
