+++ src/main/kotlin/mashup/spring/elegant/search/domain/search/factory/impl/FunctionalSingleTermQueryFactory.kt
@@ -1,11 +1,11 @@
 package mashup.spring.elegant.search.domain.search.factory.impl
 
 import mashup.spring.elegant.search.domain.search.addLocationCondition
-import mashup.spring.elegant.search.domain.search.addRequiredConditions
+import mashup.spring.elegant.search.domain.search.addAcceptableConditions
 import mashup.spring.elegant.search.domain.search.enums.SearchType
 import mashup.spring.elegant.search.domain.search.factory.FunctionQueryFactory
 import mashup.spring.elegant.search.domain.search.factory.QueryBaseFactory
-import mashup.spring.elegant.search.domain.search.factory.TypeQueryFactory
+import mashup.spring.elegant.search.domain.search.factory.SingleTermQueryFactory
 import mashup.spring.elegant.search.dto.SearchDto
 import org.elasticsearch.index.query.QueryBuilder
 import org.springframework.stereotype.Component
@@ -14,11 +14,11 @@ import org.springframework.stereotype.Component
  * Abstract Factory Impl
  */
 @Component
-class FunctionTypeQueryFactory (
+class FunctionalSingleTermQueryFactory (
     private val queryFactory : QueryBaseFactory,
     private val featureFactory : MustFeatureQueryFactory,
     private val functionFactory : FunctionQueryFactory
-): TypeQueryFactory{
+): SingleTermQueryFactory{
 
 
     override fun build(type: SearchType, dto: SearchDto.Request): QueryBuilder
@@ -27,7 +27,7 @@ class FunctionTypeQueryFactory (
 
                 val boolQuery = queryFactory.create(type, dto.term)
                                                             .addLocationCondition(dto.lat, dto.lon)
-                                                            .addRequiredConditions(dto.area)
+                                                            .addAcceptableConditions(dto.area)
 
                 val featureQuery = featureFactory.create(type, dto.feature, boolQuery)
