+++ src/main/kotlin/mashup/spring/elegant/search/domain/search/mapper/IdScoreMapper.kt
@@ -1,7 +1,7 @@
 package mashup.spring.elegant.search.domain.search.mapper
 
 
-import mashup.spring.elegant.search.domain.search.Shop
+import mashup.spring.elegant.search.domain.model.Shop
 import mashup.spring.elegant.search.dto.SearchDto
 import org.springframework.data.elasticsearch.core.SearchHits
 import org.springframework.stereotype.Component
